+++ plugins/events-backend/src/index.ts
@@ -23,3 +23,4 @@
 export { EventsBackend } from './service/EventsBackend';
 export { InMemoryEventBroker } from './service/InMemoryEventBroker';
 export { HttpPostIngressEventPublisher } from './service/http';
+export { DefaultEventBroker } from './service/DefaultEventBroker';