+++ plugins/kubernetes-backend/src/service/KubernetesFanOutHandler.ts
@@ -30,7 +30,6 @@ import {
   ServiceLocatorRequestContext,
 } from '../types/types';
 import { KubernetesAuthTranslator } from '../kubernetes-auth-translator/types';
-import { KubernetesAuthTranslatorGenerator } from '../kubernetes-auth-translator/KubernetesAuthTranslatorGenerator';
 import {
   ClientContainerStatus,
   ClientCurrentResourceUsage,
@@ -137,7 +136,9 @@ export const DEFAULT_OBJECTS: ObjectToFetch[] = [
 ];
 
 export interface KubernetesFanOutHandlerOptions
-  extends KubernetesObjectsProviderOptions {}
+  extends KubernetesObjectsProviderOptions {
+  authTranslator: KubernetesAuthTranslator;
+}
 
 export interface KubernetesRequestBody extends ObjectsByEntityRequest {}
 
@@ -195,7 +196,7 @@ export class KubernetesFanOutHandler {
   private readonly serviceLocator: KubernetesServiceLocator;
   private readonly customResources: CustomResource[];
   private readonly objectTypesToFetch: Set<ObjectToFetch>;
-  private readonly authTranslators: Record<string, KubernetesAuthTranslator>;
+  private readonly authTranslator: KubernetesAuthTranslator;
 
   constructor({
     logger,
@@ -203,13 +204,14 @@ export class KubernetesFanOutHandler {
     serviceLocator,
     customResources,
     objectTypesToFetch = DEFAULT_OBJECTS,
+    authTranslator,
   }: KubernetesFanOutHandlerOptions) {
     this.logger = logger;
     this.fetcher = fetcher;
     this.serviceLocator = serviceLocator;
     this.customResources = customResources;
     this.objectTypesToFetch = new Set(objectTypesToFetch);
-    this.authTranslators = {};
+    this.authTranslator = authTranslator;
   }
 
   async getCustomResourcesByEntity({
@@ -313,12 +315,7 @@ export class KubernetesFanOutHandler {
     // Execute all of these async actions simultaneously/without blocking sequentially as no common object is modified by them
     const promiseResults = await Promise.allSettled(
       clusterDetails.map(cd => {
-        const kubernetesAuthTranslator: KubernetesAuthTranslator =
-          this.getAuthTranslator(cd.authProvider);
-        return kubernetesAuthTranslator.decorateClusterDetailsWithAuth(
-          cd,
-          auth,
-        );
+        return this.authTranslator.decorateClusterDetailsWithAuth(cd, auth);
       }),
     );
 
@@ -392,20 +389,5 @@ export class KubernetesFanOutHandler {
 
     result.errors.push(...podMetrics.errors);
     return [result, podMetrics.responses as PodStatusFetchResponse[]];
-  }
-
-  private getAuthTranslator(provider: string): KubernetesAuthTranslator {
-    if (this.authTranslators[provider]) {
-      return this.authTranslators[provider];
-    }
-
-    this.authTranslators[provider] =
-      KubernetesAuthTranslatorGenerator.getKubernetesAuthTranslatorInstance(
-        provider,
-        {
-          logger: this.logger,
-        },
-      );
-    return this.authTranslators[provider];
   }
 }