+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/test-utils
 
+## 1.3.0-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/config@1.0.7
+  - @backstage/core-app-api@1.7.0-next.2
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/types@1.0.2
+  - @backstage/plugin-permission-common@0.7.5-next.0
+  - @backstage/plugin-permission-react@0.4.12-next.1
+
 ## 1.3.0-next.1
 
 ### Patch Changes