+++ plugins/gcp-projects/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-gcp-projects
 
+## 0.3.36-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.6-next.2
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/theme@0.2.19-next.0
+
 ## 0.3.36-next.1
 
 ### Patch Changes