+++ plugins/periskop-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-periskop-backend
 
+## 0.1.15-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.4-next.2
+  - @backstage/backend-plugin-api@0.5.1-next.2
+  - @backstage/config@1.0.7
+
 ## 0.1.15-next.1
 
 ### Patch Changes