+++ plugins/tech-insights/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-tech-insights
 
+## 0.3.9-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.6-next.2
+  - @backstage/plugin-catalog-react@1.4.1-next.2
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/catalog-model@1.2.1
+  - @backstage/errors@1.1.5
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/types@1.0.2
+  - @backstage/plugin-tech-insights-common@0.2.10
+
 ## 0.3.9-next.1
 
 ### Patch Changes